requirements.txt
Flask
Flask-Cors
websocket-client
Flask-SQLAlchemy
Flask-Login
requests

